Move ExampleApplication story to root storybook workspace
Avoids a circular devDependency between @wordpress/theme and
@wordpress/ui. Also excludes storybook stories from the
use-recommended-components lint rule.
