+++ packages/theme/package.json
@@ -92,7 +92,6 @@
 		"@testing-library/jest-dom": "^6.9.1",
 		"@types/jest": "^29.5.14",
 		"@types/node": "^20.17.10",
-		"@wordpress/ui": "file:../ui",
 		"esbuild": "^0.27.2",
 		"esbuild-esm-loader": "^0.3.3",
 		"storybook": "^10.1.11"
